Several major changes in the scripts:
1- All the scripts have been altered to work exclusively with Gadget's
format 2 (set by ICFormat 2 and SnapFormat 2). This format, among other
advantages, makes it easier to handle optional blocks, in case they
are implemented in the future.

2- The interface between cluster.py and snapwrite.py is no longer made
through text files, but directly in the act of calling the funcion
write_snapshot, by means of an adicional argument contaning all the
necessary data.

3- A major bug in the rejection algorithm has been fixed (although not
identified), and now the radial velocity distribution is in accordance
with the values calculated analytically.

4- All the rejection algorithm for sampling the velocities has been moved
to Cython, including the auxiliary funcion to calculate the potential at
a given radius. The performance has improved a lot.

5- analysis.py has been configured to work even without a X session - e.g.
under an SSH session.