Algorithm to find the COD implemented
This algorithm, which finds the "center" of the cluster, is necessary,
in order to measure the radial distributions of interest appropriately.

This is achieved by means of an interactive algoritm, that starts
at the center of mass, and moves in the direction of the potential
well at first with large steps, and then at increasignly smaller
steps, until the desired precision is achieved.