Gas capability added to the scripts.
Now the script cluster.py is capable of generating a galaxy cluster halo
containing not only dark matter, but dark matter and gas. The gas, as the
dark matter did before, follows a Hernquist density profile.

Also, the analysis script analysis.py, which has been named to profiles.py,
now detects automatically whether the snapshot contains gas or not, and, if
it does, generates an additional temperature distribution plot.